+++ contracts/ybr/EthYBR.sol
@@ -35,7 +35,7 @@ import { Time } from "@openzeppelin/contracts/utils/types/Time.sol";
  * @notice This contract is for the YieldBricks token, which is an ERC20 token with additional features.
  */
 contract EthYBR is YBRBase, ERC20L1BridgeableUpgradeable {
-    uint private constant _CAP = 1_000_000_000 ether;
+    uint256 private constant _CAP = 1_000_000_000 ether;
 
     // /// @notice Contract constructor - disabled due to upgradeability
     // /// @custom:oz-upgrades-unsafe-allow constructor
